+++ DynatracePS/Private/Invoke-DynatraceAPIMethod.ps1
@@ -66,10 +66,18 @@ function Invoke-DynatraceAPIMethod {
         }
 #endregion Headers
 
+        # Create a hash table from the query in the URI, makes it easier to 
+        # manage in powershell
+        #
         $uriQuery = ConvertTo-ParameterHash -Uri $Uri
 
+        # Generate the Get parameter hashtable that we will use
+        #
         $internalGetParameter = Join-Hashtable $uriQuery, $GetParameter
         [Uri]$LeftPartOfURI = $Uri.GetLeftPart('Path')
+
+        # Concat the url and query together to form the URI address we are going to use
+        #
         [Uri]$FinalURI = "{0}{1}" -f $LeftPartOfURI,(ConvertTo-GetParameter $internalGetParameter)
 
         Write-Verbose "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] FinalURI: [$FinalUri]"
@@ -80,12 +88,16 @@ function Invoke-DynatraceAPIMethod {
                 Method = $Method
                 Headers = $_headers
             }
+            # If -body parm is used, we add it to the splat parameters
+            #
             if ($body) {
                 Write-Debug "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] body: $($body | Out-String)"
                 $splatParameters += @{
                     Body = $body
                 }
             }
+            # if contenttype is defined, add it to the parameters
+            #
             if ($ContentType) {
                 $splatParameters += @{
                     ContentType = $ContentType
@@ -94,6 +106,12 @@ function Invoke-DynatraceAPIMethod {
 
             Write-Debug "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] splatParameters: $($splatParameters | Out-String)"
 
+            # Invoke-WebRequest (IWR) to Dynatrace. We use IWR instead of Invoke-RestMethod (IRM) because of reasons:
+            # 1) IWR is standard from PS version 3 and up. IRM is not
+            # 2) IRM doesn't do good job of returning headers and status codes consistently. IWR does.
+            #
+            # https://www.truesec.com/hub/blog/invoke-webrequest-or-invoke-restmethod
+            #
             $Response = Invoke-WebRequest @splatParameters
             $RestResponse = $Response.Content | ConvertFrom-JSON
             $ResponseHeaders = $Response.Headers
@@ -110,6 +128,8 @@ function Invoke-DynatraceAPIMethod {
         }
         Write-Verbose "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] Executed RestMethod"
 
+        # If there are bad status codes, this will break and cause function to exit
+        #
         Test-ServerResponse -InputObject $RestResponse -StatusCode $StatusCode
     }
 
@@ -117,6 +137,8 @@ function Invoke-DynatraceAPIMethod {
         if ($RestResponse) {
             Write-Verbose "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] nextPageKey: $($RestResponse.nextPageKey)"
 
+            # Set next page key to get
+            #
             $NextPageKey = $RestResponse.nextPageKey
 
             if ($RestResponseProperty) {
@@ -132,10 +154,12 @@ function Invoke-DynatraceAPIMethod {
             do {
                 
                 if (-not $NextPageKey) {
-                    # if there is no page key, then quit
+                    # if there is no page key, then quit, as there are no more pages
+                    #
                     break
                 } else {
-                    # Output results from this loop
+                    # Output results from this loop, and continue on the recursion
+                    #
                     $result
                 }
                 
@@ -143,6 +167,9 @@ function Invoke-DynatraceAPIMethod {
                 $PSBoundParameters["LevelOfRecursion"] = $LevelOfRecursion + 1
 
                 Write-Verbose "[$($MyInvocation.MyCommand.Name) $LevelOfRecursion] Calling Invoke-DynatraceApiMethod"
+
+                # Call this same function again (recursion)
+                #
                 $result = Invoke-DynatraceApiMethod @PSBoundParameters
             } while (-not $NextPageKey)