+++ Changelog.md
@@ -6,15 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+<<<<<<< HEAD
 ## [0.1.15]
 ### Added
 -Function Get-DynatraceHTTPCheck to get all the http checks in dynatrace
 -Function Get-DynatraceHTTPCheckProperty to get properties for an http check
 
 ## [0.1.14]
+=======
+## [0.1.14] - 2022-06-10
+>>>>>>> ad5e3079fdc314e44837dec6ab1f316e5e06eaa2
 ### Added
 -Function Get-DynatraceManagementZoneRule to get all the rules associated with a management zone
 -Function Get-DynatraceHostsInHostGroup to return a list of hosts and related host groups
+
+## [0.1.13] - 2022-06-10
+### Added
+-Get-DynatraceProblem to get problems
+-Internal function ConvertTo-Datetime to convert millisecond timestamps used in dynatrace to human readable local time
 
 ## [0.1.12]
 ### Added